+++ README.md
@@ -1,7 +1,7 @@
 # python-gui-template
 
-Replace "MyApp" anywhere in the code by the name of your package, such as:
+Replace "pyguitemp" anywhere in the code by the name of your package, such as:
 
 - setup.cfg (x2)
-- the myapp folder itself
+- the pyguitemp folder itself
 - docs/source/conf.py
