--- myapp/__init__.py
@@ -1,2 +1,0 @@
-__VERSION__ = "0.0.1"
-APP_NAME = "MyApp"
