+++ pyguitemp/__init__.py
@@ -1,0 +1,2 @@
+__VERSION__ = "0.0.1"
+APP_NAME = "pyguitemp"
