+++ pyguitemp/extensions/about_dialog.py
