+++ setup.cfg
@@ -3,14 +3,14 @@ requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [metadata]
-name = myapp
+name = Python GUI Template
 version = 0.0.1
 url = https://github.com/ImperialCollegeLondon/python-gui-template
 author = Research Computing Service, Imperial College London
 author_email = rcs-support@imperial.ac.uk
 
 [options]
-packages = myapp
+packages = pyguitemp
 install_requires =
     wxPython
 
@@ -45,8 +45,8 @@ test = pytest
 [tool:pytest]
 addopts =
 	-v --flake8 --mypy -p no:warnings
-	--cov=myapp  --cov-report=html:htmlcov/coverage
-	--doctest-modules --ignore=myapp/__main__.py
+	--cov=pyguitemp  --cov-report=html:htmlcov/coverage
+	--doctest-modules --ignore=pyguitemp/__main__.py
 
 [pycodestyle]
 max-line-length = 88
